Fix bug in Connections with errors
Summary:
This diff fixes a bug in the connection handler where errors that occur on a field are not copied into the "connection" record for that field. Connections create virtual fields on a record that contain the results for the set of edges fetched with a set of args. These are merged into the record by the connection handler. However, the errors are only ever set on the base field in the normalizer. By copying the fields over to the connection field, the `throwOnFieldError` directive now works correctly on connections.

This diff also adds some public APIs to the RecordProxy and mutator to deal with errors, which have not been handled in Relay previously.

Reviewed By: captbaritone

Differential Revision: D64505652

fbshipit-source-id: 288b33bf047d5baf7dacd0961ba6b5692766be0e