Update normalization artifacts for exec time resolver strong ids
Summary: This diff updates artifact generation for exec time resolvers to include the resolver module in the normalization AST. It's fairly similar to how the reader AST is built but we need far less information at normalization time. This is the most basic version of the artifact and we may need to add additional fields as the feature continues to be built.

Reviewed By: captbaritone

Differential Revision: D64635272

fbshipit-source-id: 3627fbb51570e4559bf98a0334b136d11d0f0de1