Fix publishing npm packages (#4833)
Summary:
This probably blames back to https://github.com/facebook/relay/issues/4816 but I haven't completely proven that.

Essentially, this looks likely that something changed in npm. `npm-package-arg` thinks that `dist/foo` is a reference to a github repo, not a local path. Explicitly making that a relative path (`./dist/foo`) ensures it doesn't.

I haven't proven this will fix the issue but running `require('npm-package-arg')('dist/foo')` returns an object indicating it's treating it as a git repo. `require('npm-package-arg')('./dist/foo')` treats it like the directory it is.

My assumption is that this treatment like a git repo is that npm is deciding to go look it up and validate that the `main` tag exists. That might be fine but it's looking for `dist/foo` which doesn't exist and then that results in the git failure we're seeing.

Pull Request resolved: https://github.com/facebook/relay/pull/4833

Reviewed By: captbaritone

Differential Revision: D65003728

Pulled By: lynnshaoyu

fbshipit-source-id: bad04bbca99cb7cfc028fcc4d74c9f378df689a8