Update connections-pagination.md
I found out that `useTransition` doesn't work when using `usePaginationFragement`. To get it to work the way it's intended in the tutorial I had to use `isLoadingNext` boolean prop that is returned from `usePaginationFragement` instead. Not sure if this a bug, or by design.

As a beginner, I spent quite some time trying to figure out why my code wasn't working, after a bit of digging I ran into some potenially related issues:
- https://github.com/facebook/relay/issues/4531#issuecomment-1867831999
- https://github.com/facebook/relay/issues/4526
- https://github.com/facebook/relay/issues/3082

In this PR I've updated the "Improving the Loading Experience with useTransition" example not to confuse newcomers.