Add support for @catch on fragments/queries/mutations and aliased inline fragments (#4838)
Summary:
`catch` was originally conceived as a means for handling field errors, so we built it as a directive on fields. However, as the design evolved we opted to make it additionally catch any errors nested within it. With this newer approach, there’s no reason to limit `catch` to just fields. Any GraphQL construct which gets read as a concrete value should be able to operate as a `catch` boundary.

This PR adds support for marking fields/queries/mutations and aliased fragments as `catch` boundaries. This means:

1. Their value will be read out as either nullable or a `Result` type
2. `semanticNonNull` values nested within the fragment/operation/mutation/aliased fragment will be typed using their semantic type (non-nullable)

One particularly nice thing this enables is that it allows `catch` to act as a non-destructive alternative to `throwOnFieldError`.

## Additional Bug Fixes

Along the way I noticed and fixed a few bugs:

1. Semantic nullability was not respected within inline fragments nested within `catch` or `throwOnFieldError`
2. Errors that bubbled up to a `catch(to: NULL)` directive would not actually cause the value to become null.

Pull Request resolved: https://github.com/facebook/relay/pull/4838

Reviewed By: itamark

Differential Revision: D65432323

Pulled By: captbaritone

fbshipit-source-id: 7fd1313e6d1ae0bda2d6354ef931f10fcfa72e9e