Support dynamic switching of read vs exec time resolvers in reader
Summary:
This diff add runtime support for swapping between exec-time and read-time resolvers in the reader. It adds a flag to the operation in the artifact which indicates if the query is an exec-time query or not. This flag is then looked up in the reader and, if it is found, the reader ignores the resolver fields and treats the ReaderAST nodes as standard nodes (either using a standard scalar field in the case of a scalar resolver or a linked field in the case of a strong object edge resolver).

Note that this does not fix the issue of a single resolver being used in both read-time and exec-time queries. This will be addressed in a follow up diff.

Reviewed By: captbaritone

Differential Revision: D65639792

fbshipit-source-id: 68a13f9a2f8d668af5fee5863e12269261df2ba2