Move `get_all_roots` from WatchmanFileSource to Config
Moving the `get_all_roots` into `Config` so it can be shared between
`WatchmanFileSource` and `WalkDirFileSource`. This returns a refined list of all
the roots within the top root directory to limit which folders are searched, but
the optimization is only used for Watchman and not for the directory walking
implementation in WalkDirFileSource.