Limit WalkDir to the root directories referenced by the configuration (#4850)
Summary:
When not using Watchman the Relay Compiler will iterate over all files under `root_dir`, which can possibly include the entire repository. This is unnecessary work and might slow down large repositories.

The watchman implementation only listens to files under the roots defined by the config, but this is not implemented for the directory walking implementation in `WalkDirFileSource`.  This change moves the `get_all_roots()` functionality from `WatchmanFileSource` to `Config` and shares it with `WalkDirFileSource` to make it only iterate over the folders configured as targets in Relay.

This fixes https://github.com/facebook/relay/issues/4849

Pull Request resolved: https://github.com/facebook/relay/pull/4850

Reviewed By: tyao1

Differential Revision: D66390485

Pulled By: captbaritone

fbshipit-source-id: cdd6daa4368419dfcc012e50f8be45440f02cdcf