Add docs for @gqlField resolvers
Summary: This diff adds some docs for the current state of the simplified syntax for property lookup resolvers. We may want to expand this into its own page in the future if we expand this to have more options and/or work on strong types but I think it fits in the current doc on defining fields.

Reviewed By: captbaritone

Differential Revision: D67425248

fbshipit-source-id: a4de39ad03b25f2cccd16efae6c5c365862941a1