Add write_arrow_fn to the printer
Summary: This is a small step in the right direction of making the printer use node printers closer to the JS AST. I thought it would be a bit more useful and recieve other AST nodes as the parameters and body, but I think that would be a larger refactor. This is just a first step towards a better pattern.

Reviewed By: captbaritone

Differential Revision: D67406271

fbshipit-source-id: f792a329edface099dbc5e6113e9dd577b4c1c6a