Add plural fragment support to `observeFragment()` (#4862)
Summary:
This PR adds plural fragment support to `observeFragment()` (and `waitForFragmentData()` which uses `observeFragment()` under the hood)

It should be noted that the Relay environment currently doesn't support subscribing to plural fragment updates, so the implementation is done by subscribing to each selector and merging the states. Since the current implementation doesn't do any batching of notifications that are made from each selector subscription, it can be somewhat inefficient if the consumer doesn't batch on their own.

Pull Request resolved: https://github.com/facebook/relay/pull/4862

Reviewed By: tyao1

Differential Revision: D67148721

Pulled By: captbaritone

fbshipit-source-id: 1c63178c78ece3588ac86b8814d9b8fbf2ce057b