Unblock Cargo.lock update workflow (#4887)
Summary:
The "Update Cargo.lock" GitHub workflow has been stuck for about a month. It's now making the GitHub CI red since Cargo.lock needs to be updated.
A lint suppression for an unknown lint is blocking the compilation of the common crate. The lint suppression was added to get Relay to build with a new Rust toolchain version. So, one possibility is that we need to bump the Rust toolchain used in the Cargo.lock action to match what we have in the other GitHub actions.

Pull Request resolved: https://github.com/facebook/relay/pull/4887

Reviewed By: captbaritone

Differential Revision: D68478053

Pulled By: monicatang

fbshipit-source-id: c0b7abd6179b761a88c19ba9e8ab6b2acce66b5f