Update docs on usePaginationFragement & transitions in Relay tutorial. (#4842)
Summary:
I found out that `useTransition` doesn't work when using `usePaginationFragement`. To get it to work the way it's intended in the tutorial I had to use `isLoadingNext` boolean prop that is returned from `usePaginationFragement` instead. Not sure if this a bug, or by design.

As a beginner, I spent quite some time trying to figure out why my code wasn't working, after a bit of digging I ran into some potenially related issues:
- https://github.com/facebook/relay/issues/4531#issuecomment-1867831999
- https://github.com/facebook/relay/issues/4526
- https://github.com/facebook/relay/issues/3082

In this PR I've updated the code in "Improving the Loading Experience with useTransition" example not to confuse newcomers. Also the section about "results that aren’t immediate should be wrapped in a React transition" as it no longer applies in this particular context.

I also replaced `<CommentsLoadingSpinner />` (which doesn't exist?) with `<SmallSpinner />` which seems more appropriate in the particular example.

Pull Request resolved: https://github.com/facebook/relay/pull/4842

Reviewed By: tyao1

Differential Revision: D66390915

Pulled By: captbaritone

fbshipit-source-id: b8866932c627dc0d067f60008e1272dd4257013a