Avoid calling fail in tests
Summary:
Looks like we call fail() here which has special meaning in Jest tests but newer versions of Jest don't implement it, and our OSS test run in that newer version.
https://github.com/jestjs/jest/issues/11698

Trying to see if it's possible to remove that call to fail() so that OSS and internal tests have consistent behavior.

Reviewed By: itamark

Differential Revision: D68867479

fbshipit-source-id: 896ffff042849dfff6fe7ff88d61084e20d369dd