Error on empty selections after fragment argument transform (#4908)
Summary:
One of my colleagues contacted me today, because he received the following error from the compiler:
> thread 'main' panicked at crates/relay-compiler/src/build_project/generate_artifacts.rs:153:13:
Expected at least one of an updatable reader AST, or normalization AST to be present

After some debugging I found out this happens because an `include` was removing all selections on a linked field, because it received a `false` constant through fragment arguments from a test operation.

This PR updates the fragment argument transform to report an appropriate error earlier, outlining the cause of the problem.

Pull Request resolved: https://github.com/facebook/relay/pull/4908

Reviewed By: captbaritone, itamark

Differential Revision: D69320183

Pulled By: lynnshaoyu

fbshipit-source-id: 26cc4f913d905e3167c0a0953075be9398fc34ce