Remove bundler setup (#4935)
Summary:
This PR removes the bundler setup for all of the npm packages in this repo, as discussed with captbaritone in [Discord](https://discord.com/channels/625400653321076807/1342499490011025460/1346175975628734494).
While this wouldn't affect the vast majority of use cases, I believe it should technically be considered a breaking change, as it breaks CDN usages and direct CJS requires pointing at outputs.

Pull Request resolved: https://github.com/facebook/relay/pull/4935

Reviewed By: tyao1

Differential Revision: D70794723

Pulled By: captbaritone

fbshipit-source-id: a4110cf14da26d81995f89a55e3b19c1c2abab32