Add `disablePrefetching` prop to `usePrefetchableForwardPaginationFragment`
Summary:
There are times when we want `usePrefetchableForwardPaginationFragment` to stop prefetching (e.g. when an error occurs), but we currently have no way to do so. This adds a prop to disable prefetching if desired.

It's worth noting that it only disables the prefetch -- `loadMore` will still call the query, so it's up to the parent component to stop calling `loadMore` when necessary.

Reviewed By: tyao1

Differential Revision: D71504733

fbshipit-source-id: 68d340aa96d01406a5ed382625518b8947d20a4c