relay-config,relay-compiler: add relativizeJsModulePath option
All CommonJS modules are currently forced to be relative to the file
they were referenced in. This makes some runtime JS integrations
impossible, e.g. reusing an NPM package's default export as a provider
(since `provider: "@example/foo"` is emitted as
`provider: require("./@example/foo")`). Add a new config setting to allow
consumers to opt out of this automatic ./ prefix.

NOTE: The new setting is not respected anywhere in this commit; it's
only declared.

supports #4456