relay-transforms: optionally use native gql fragment spreads for unmasked fragments
When Relay encountered a fragment spread with the `@relay(mask: false)`
directive, it would previously inline that fragment text into the query
itself. This is completely valid and equivalent to a native GraphQL
fragment spread. That inlining can dramatically increase the size of
serialize for a query with multiple fragments. Such repetition compresses
extremely well, but clients and servers still pay the cost of having
uncompressed strings in-memory (plus compress/decompress time).

Allow relay users to opt out of fragment inlining for
`@relay(mask: false)` fragment spreads in queries, using a native
GraphQL fragment spread. The reader AST and type generation are not
affected so there's no impact to the relay runtime.