Handle relay:// json schema requests even outside of Relay projects (#4961)
Summary:
Previously if you opened a `package.json` file in a non-Relay projects you would see an error like this:

![Screenshot 2025-04-22 at 4 36 36 PM](https://github.com/user-attachments/assets/0625fce8-a25c-49bc-912d-52d1f2c95784)

This was because we didn't register a handler if no Relay project was detected. Now we register a noop handler, and there is no error.

![Screenshot 2025-04-22 at 4 32 28 PM](https://github.com/user-attachments/assets/011037cb-b349-4b41-8002-a2f6d8b59482)

Thanks to cpojer for the report.

Pull Request resolved: https://github.com/facebook/relay/pull/4961

Reviewed By: lynnshaoyu

Differential Revision: D73476115

Pulled By: captbaritone

fbshipit-source-id: d093cb97ea7797f382be65d2ba4d535aa00043c5