add missing generic type parameter for mutation tutorial example (#4981)
Summary:
https://relay.dev/docs/tutorial/mutations-updates/#step-4--modify-the-updatable-data

This step of the tutorial is currently giving us type errors.

![CleanShot 2025-05-05 at 18 09 58@2x](https://github.com/user-attachments/assets/02359c4b-5910-4c81-9b12-2d914dbd7eca)

This is because we are missing the generic type parameter.

Note that this will also require us to upstream the relay dependencies in the example to the latest v19: https://github.com/relayjs/relay-examples/pull/333

After that, our tutorial will be working properly with no type error!

Pull Request resolved: https://github.com/facebook/relay/pull/4981

Reviewed By: evanyeung

Differential Revision: D74219468

Pulled By: captbaritone

fbshipit-source-id: f2c37de0d0400e0a376cdd7ca89c1c7ef3bd0946