Add page documenting the compiler config (#4985)
Summary:
Our compiler config is relatively huge and small details of it change frequently. Keeping it documented and up to date has proved too difficult. However, recently we started driving a JSON schema representation of it and shipping that as part of the VSCode extension to give in-editor feedback about what options exist and to catch errors.

However there still has not been an exhaustive place to see what options exist or read about them.

This PR attempts to solve that by adding a new documentation page with a hand-written React renderer that traverses the JSON schema and renders a friendly version of it that looks ~like TypeScript types, which should be familiar to our users.

Some problems to solve if we want to go this way:

- [x] How should we sync the config schema file to the website folder?
- [x] How should we automate testing that schema changes do not break this?
  - I've confirmed that Docusaurus build will fail if we throw in the React components. So, if we are okay with that being our "test" that we can correctly render our config docs, we can just ensure that our react components are defensive and throw on any unhandled cases.
- [x] Interlink (and remove duplication) with existing compiler page
- [x] See if we can fix anchor links (it scrolls such that the anchor ends up under the header)
- [x] Handle dark mode

![localhost_3000_docs_next_getting-started_compiler-config_ (1)](https://github.com/user-attachments/assets/fce7eb4e-bb54-4685-bc73-eb96bbd4109e)

Pull Request resolved: https://github.com/facebook/relay/pull/4985

Reviewed By: evanyeung

Differential Revision: D74346512

Pulled By: captbaritone

fbshipit-source-id: e4443df73bbe01631c7c7b7ec915f81d541d0fc5