Fix variable name for imported model resolvers when using ES module imports (#4984)
Summary:
Model resolvers are currently broken in v19.0.0 when used with ES modules. We were reusing the local variable name for the parent resolver which meant we got a name collision and we would try to use the parent field's resolver to resolve the model rusting in a runtime error. This is an initial attempt to remedy the issue.

Pull Request resolved: https://github.com/facebook/relay/pull/4984

Test Plan:
Tried this on a test project before importing and saw that it allowed a test project with resolvers to run.

Ran on WWW and saw that it did not produce any Flow errors. (D74268167)

## AM Labs Before and After (identical output

 {F1977878641}

 {F1977878642}

Reviewed By: tyao1

Differential Revision: D74261741

Pulled By: captbaritone

fbshipit-source-id: c738eea0188f8fc218bec22f830e483af02e0f37