Emit raw query text in artifact for `@relay_test_operation` operations (#4999)
Summary:
Hello!

In Microsoft we have a use case where we would like to parse Relay operations being executed with `graphql-js`. In our case this is specifically for testing with Storybook, where we want to use [Graphitation payload generator](https://github.com/microsoft/graphitation/tree/main/packages/graphql-js-operation-payload-generator) to generate mock data, which takes `graphql-js` operations as input.

Today, parsing operations is somewhat limited because the `text` field in Relay operation artifacts is heavily transformed and for example does not contain client extensions. We would like to expose the raw, untransformed query text in operation artifacts, hidden behind the `relay_test_operation` directive. We believe this would be useful in other testing scenarios as well, because it enables Relay operations to be easily converted between formats and used with other GraphQL tooling.

This PR implements that by adding a `rawText` field in `OperationDefinition` and a new transform, the `RawTextTransform`. The transform populates this field, only on operations with `relay_test_operation` directive. The transform is specifically added as the first transform to be executed, since it relies on generating the original, "raw" query text, before any other transforms are applied.

Here's an example on how `rawText` would be exposed in an artifact:

<img width="1898" alt="image" src="https://github.com/user-attachments/assets/1d5ed008-98b0-42ae-a7a3-c5198383dc08" />

where you can see that `viewData`, which is a client-extension, is preserved.

I've added this field also to Flow types, and if you think it makes sense to mention this in docs, I can add that as well.

Would love to hear if you have any thoughts on adding this feature :)

Pull Request resolved: https://github.com/facebook/relay/pull/4999

Reviewed By: alunyov

Differential Revision: D75008104

Pulled By: tyao1

fbshipit-source-id: 45a9150b21f51500fa1c499780d412f4e387d4c0