config: allow 'extra' in root of single-project config files
The 'extra' property was previously only declared for the projects in a
multi-project config file. Since unknown fields are considered an error
when parsing single-project config files, this made it impossible to
include arbitrary metadata via a package.json config or a
relay.config.json file. Allow 'extra' at the root of a single-project
config file to better-match the structure of projects in a multi-project
config.

fixes #5029