Update schema-flatbuffer with current version of flatbuffer - v2
Summary:
Original diff caused a problem because the new validation system in flatbuffers to prevent too many tables doesn't work with us... so changed it to use the "unverified" method.

Description of the original diff (which was reverted).

We had flatbuffers-old (with is vulnerable) for ages because of a few things that had their _generated.rs file generated by very old version of flatc...
I just downloaded a fresh flatc and generated the file with it - -links just fine with the current flatbuffers, no need to keep using the old one.
 Regenerate graphqlschema_generated.rs with flatbuffer 25.2.10

Reviewed By: mjmahone

Differential Revision: D78011651

fbshipit-source-id: 5b166ac925822e339344f16bea568fb9aa177d9a