rmcp 0.2.1 -> 0.3.0, schemars 0.8.21 -> 1.0.4 (#5035)
Summary:
Pull Request resolved: https://github.com/facebook/relay/pull/5035

Update rmcp (the Model Context Protocol crate) from 0.2.1 to 0.3.0. Not a lot of breaking changes this time, except that it imports schemars which causes a collision in testcompose which also uses it, but the old version we had around... so this diff also updates schemars and it's an API breaking update.

Let's see what happens with CI.

Schemars: Followed the migration guide the best I could: https://graham.cool/schemars/migrating/

But fingers crossed for tests being good.

Warning: As a result of the change in schemars **the output has changed and I've "fixed" the tests to match the new output!**.

Some changes are expected (I think), for example:

```
-  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
```

Seems reasonable.

Other changes such as new line breaks like this:
```
-  "description": "Relay's configuration file. Supports a single project config for simple use cases and a multi-project config for cases where multiple p
rojects live in the same repository.\n\nIn general, start with the SingleProjectConfigFile.",
+  "description": "Relay's configuration file. Supports a single project config for simple use\ncases and a multi-project config for cases where multiple
projects live in\nthe same repository.\n\nIn general, start with the SingleProjectConfigFile.",
```
might cause problems?

And this one:
```
-      "allOf": [
-        {
-          "$ref": "#/definitions/SingleProjectConfigFile"
-        }
-      ]
+      "$ref": "#/$defs/SingleProjectConfigFile"
```

I have no idea if it's correct.

 Update schemars crate from 0.8 to 1.0.4

Reviewed By: Imxset21, diliop

Differential Revision: D78455871

fbshipit-source-id: b041e56797db9c5b784a292913c190fba050a54f