Make schema def rendering more robust (#5036)
Summary:
In https://github.com/facebook/relay/commit/b208fedd1c1e86cc63f69e9dfbe2accf6b240283 we upgraded schemars which changed how the json schema for the compiler config is represented. This uncovered gaps in our documentation renderer. Specifically, we had hard coded the assumption that the definitions would live on the `.defintitions` property of the schema. The new version of schemars puts them under `$defs` which aligns better with json schema spec.

This diff makes `$defs` the default assumed location for definitions, but lets the parent pass in definitions from another location. This diff also starts the process of using versioned schemas for our versioned docs. As part of this I pulled in the version of the config schema from the commit that added the v20 docs. So, now as our current config schema changes, the v20 docs will still reflect the config as it was at the v20 release.

This allows us to pressure test the docs rendering, ensuring they can render both the old and new representations.

Pull Request resolved: https://github.com/facebook/relay/pull/5036

Reviewed By: cfsmp3

Differential Revision: D78595366

Pulled By: captbaritone

fbshipit-source-id: 589b320847859b9196b82e473669026d0c899a4b