Update broken url in error message
Summary:
This diff updates the reference url in the error message for the `InvalidConnectionKeyArgPostfix` error.

Also fix some linter errors that were picked up by CI 😅

## More Details

When trying to add a new relay connection, I got an error message in the command line with a url pointing to the relay docs. Specifically, a url to `https://
relay.dev/docs/en/pagination-container#connection`.

{F1980415414}

That url leads to a page that no longer exists...

{F1980415420}

I checked the [history of this page in the wayback machine](https://web.archive.org/web/20210228095734/https://relay.dev/docs/en/pagination-container#connection). In their records, the page used to show this...

{F1980415369}

I think based off of this, the best page to replace the url with is [`https://relay.dev/docs/tutorial/connections-pagination/`](https://relay.dev/docs/tutorial/connections-pagination/).

Update the reference url in the error message for the InvalidConnectionKeyArgPostfix error.

Reviewed By: lynnshaoyu

Differential Revision: D78642574

fbshipit-source-id: eea893744c679fec73a0cea3d7c186a21883d0f6