Filter raw_response output types for concrete fields
When generating a raw_response type for a concrete fields, we may
(e.g.,) still be spreading a fragment which is for an abstract type. We
should cut down our type selections based on the actual concrete type of
the field, to generate more specific raw_response types. We should also
mark the selections has having the same concrete type as our field,
preventing redundant arms of the final TypeScript/Flow union types from
being output.

This makes the output raw response types a lot more pleasant to work
with.

Fixes #5090